Fix race condition for Copy.
Since the IsFinished was be calculated by the progress and not by the
actual copy routine if the completion was 1 then it just stopped and
cleaned up(closing files, sockets, etc.) so the next Read call would
fail and cause a race condition. Instead we just have a static var to
say when it's actually completed and that's set after io.Copy returns.